[d2d] basic line breaking
