+++ piet-test/src/picture_5.rs
@@ -10,7 +10,7 @@ pub fn draw(rc: &mut impl RenderContext) -> Result<(), Error> {
     // do something texty
     let font = rc.text().new_font_by_name("Segoe UI", 12.0).build()?;
 
-    let layout = rc.text().new_text_layout(&font, "piet text!").build()?;
+    let layout = rc.text().new_text_layout(&font, "piet text!", std::f64::INFINITY).build()?;
 
     let width = layout.width();
