+++ piet/src/text.rs
@@ -135,17 +135,17 @@ pub trait TextLayout {
 /// Metadata about each line in a text layout.
 #[derive(Debug, Default)]
 pub struct LineMetric {
+    /// Index (in code units) of the start of the line, offset from the beginning of the text.
+    pub line_start_offset: usize,
+
     /// Line length (in code units), but offset from the beginning of the text. So it's the length
     /// of this line summed with the lengths of all the lines before it.
     ///
     /// Does not include trailing whitespace.
     pub line_length_offset: usize,
 
-    /// Line width, but offset from the beginning of the text. So it's the width
-    /// of this line summed with the widths of all the lines before it.
-    ///
-    /// Does not include trailing whitespace.
-    pub line_width_offset: f64,
+//    /// Line width, does not include trailing whitespace.
+//    pub line_width: f64,
 
     /// Line length (in code units), but offset from the beginning of the text. So it's the length
     /// of this line summed with the lengths of all the lines before it.
@@ -153,11 +153,8 @@ pub struct LineMetric {
     /// Includes trailing whitespace.
     pub line_length_trailing_whitespace_offset: usize,
 
-    /// Line width, but offset from the beginning of the text. So it's the width
-    /// of this line summed with the widths of all the lines before it.
-    ///
-    /// Includes trailing whitespace.
-    pub line_width_trailing_whitespace_offset: f64,
+//    /// Line width, includes trailing whitespace.
+//    pub line_width_trailing_whitespace: f64,
 }
 
 /// return values for [`hit_test_point`](../piet/trait.TextLayout.html#tymethod.hit_test_point).
